Change speed parameter type to support more granular speeds including those less than one word per minute for QRSS use

+++ cww_MorseTx.h
@@ -14,15 +14,15 @@ void noTone(uint8_t pin);
 
 class cww_MorseTx {
   private:
-    int _dashlen;          // Length of dash
-    int _dotlen;           // Length of dot
-    int _freq;             // Tone frequency
-    bool _invert;          // true on=low else on=high  
-    byte _keypin;          // Pin for output keying
-    byte _sndpin;          // Sound pin  
+    unsigned long _dashlen; // Length of dash
+    unsigned long _dotlen;  // Length of dot
+    int _freq;              // Tone frequency
+    bool _invert;           // true on=low else on=high  
+    byte _keypin;           // Pin for output keying
+    byte _sndpin;           // Sound pin  
   public:
-    cww_MorseTx(byte keypin, byte speed, bool invert = false);
-    cww_MorseTx(byte keypin, byte speed, byte sndpin, int freq = 1000, bool invert = false);
+    cww_MorseTx(byte keypin, float speed, bool invert = false);
+    cww_MorseTx(byte keypin, float speed, byte sndpin, int freq = 1000, bool invert = false);
     void send(char c);
     void send(const char* str);
   private:
